+++ CHANGELOG.md
@@ -2,8 +2,35 @@
 
 ## Unreleased
 
-- Fix the CacheItem detection for tracing (#501)
 - Unwrap Messenger failures (`HandlerFailedException`) when using sync transport (#475)
+
+## 4.2.1 (2021-08-24)
+
+- Fix return type for `TracingDriver::getDatabase()` method (#541)
+- Avoid throwing exception from the `TraceableCacheAdapterTrait::prune()` and `TraceableCacheAdapterTrait::reset()` methods when the decorated adapter does not implement the respective interfaces (#543)
+
+## 4.2.0 (2021-08-12)
+
+- Log the bus name, receiver name and message class name as event tags when using Symfony Messenger (#492)
+- Make the transport factory configurable in the bundle's config (#504)
+- Add the `sentry_trace_meta()` Twig function to print the `sentry-trace` HTML meta tag (#510)
+- Make the list of commands for which distributed tracing is active configurable (#515)
+- Introduce `TracingDriverConnection::getWrappedConnection()` (#536)
+- Add the `logger` config option to ease setting a PSR-3 logger to debug the SDK (#538)
+- Bump requirement for DBAL tracing to `^2.13|^3`; simplify the DBAL tracing feature (#527)
+
+## 4.1.4 (2021-06-18)
+
+- Fix decoration of cache adapters inheriting parent service (#525)
+- Fix extraction of the username of the logged-in user in Symfony `5.3` (#518)
+
+## 4.1.3 (2021-05-31)
+
+- Fix missing require of the `symfony/cache-contracts` package (#506)
+
+## 4.1.2 (2021-05-17)
+
+- Fix the check of the existence of the `CacheItem` class while attempting to enable the cache instrumentation (#501)
 
 ## 4.1.1 (2021-05-10)